pyproject.toml
[project]
name = "doctor-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "connectorx>=0.4.3",
    "diff-match-patch>=20241021",
    "dspy==2.6.27",
    "ipykernel>=6.29.5",
    "mlflow>=2.22.0",
    "polars>=1.29.0",
    "pyserde>=0.24.0",
    "streamlit>=1.45.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mlflow>=2.22.0",
]

[tool.uv]
link-mode = "copy"
